Rename color header to colors, remove dep mk files
